+++ TRAP/src/Events/KeyEvent.h
@@ -53,13 +53,13 @@ namespace TRAP::Events
 		/// <param name="key">Pressed key.</param>
 		/// <param name="repeatCount">Amount of key press repeats.</param>
 		/// <param name="title">Title of the affected Window.</param>
-		KeyPressEvent(Input::Key key, uint32_t repeatCount, std::string title);
+		KeyPressEvent(Input::Key key, uint16_t repeatCount, std::string title);
 
 		/// <summary>
 		/// Retrieve tha amount of key press repeats.
 		/// </summary>
 		/// <returns>Repeat count.</returns>
-		uint32_t GetRepeatCount() const;
+		uint16_t GetRepeatCount() const;
 		/// <summary>
 		/// Retrieve the title of the affected Window.
 		/// </summary>
@@ -89,7 +89,7 @@ namespace TRAP::Events
 		const char* GetName() const override;
 
 	private:
-		uint32_t m_repeatCount;
+		uint16_t m_repeatCount;
 		std::string m_title;
 	};
